Start adding GaitControllerUnitTests

+++ ServoTest/lib/spot/GaitController.h
@@ -8,13 +8,13 @@ class GaitController {
 
     Configuration *spotConfig;
 
-    int getPhaseIndex(long ticks);
+    int getPhaseIndex(int ticks);
 
 public:
     GaitController(Configuration * config);
     ~GaitController();
-    void getFootContacts(long ticks, uint8_t *modes);
-    int getSubPhaseTicks(long ticks);
+    void getFootContacts(int ticks, uint8_t *modes);
+    int getSubPhaseTicks(int ticks);
 };
 
 #endif
